requirements.txt
python>=3.12
biopython
pytest